Lambda/audio-generation-lambda/src/requirements.txt
# Core AWS dependencies
boto3>=1.34.0

# Google AI SDK 
google-genai==1.18.0

# Supabase client 
supabase>=2.14.0

# HTTP client
requests>=2.31.0
